requirements.txt
fastapi
uvicorn
streamlit
scikit-learn
pandas
pymongo
langchain
langchain-openai
python-dotenv
joblib
numpy
pydantic
requests